Indicate that SPAT was increased from 50% to 65%
I was trying to figure out what % of staked tokens are required for a vote to win, and I mistakenly believed it was 50% when I read this page at first.  After a bit of research I learned this was increased to 65% by UMIP-186. 

I think the new 65% SPAT should be mentioned somewhere in the docs.

It's already mentioned in:
- [May 8th Update](https://docs.uma.xyz/whats-new)
- @UMAprotocol [X post](https://x.com/UMAprotocol/status/1915179067118170397)